+++ samples/Extensions/Blob/Book.cs
@@ -5,7 +5,7 @@ namespace SampleApp
 {
     public class Book
     {
-        public string? Id { get; set; }
-        public string? Name { get; set; }
+        public string Id { get; set; }
+        public string Name { get; set; }
     }
 }