+++ samples/Extensions/SignalR/SignalRTriggerFunctions.cs
@@ -10,29 +10,35 @@ namespace Extensions.SignalR
     public static class SignalRTriggerFunctions
     {
         // <snippet_on_connected>
-        [Function("OnConnected")]
-        public static void OnConnected([SignalRTrigger("chat", "connections", "connected", ConnectionStringSetting = "SignalRConnection")] SignalRInvocationContext invocationContext, FunctionContext functionContext)
+        [Function(nameof(OnConnected))]
+        public static void OnConnected(
+            [SignalRTrigger("chat", "connections", "connected", ConnectionStringSetting = "SignalRConnection")]
+                SignalRInvocationContext invocationContext, FunctionContext functionContext)
         {
-            var logger = functionContext.GetLogger("OnConnected");
-            logger.LogInformation($"Connection {invocationContext.ConnectionId} is connected");
+            var logger = functionContext.GetLogger(nameof(OnConnected));
+            logger.LogInformation("Connection {connectionId} is connected", invocationContext.ConnectionId);
         }
         // </snippet_on_connected>
 
         // <snippet_on_disconnected>
-        [Function("OnDisconnected")]
-        public static void OnDisconnected([SignalRTrigger("chat", "connections", "disconnected", ConnectionStringSetting = "SignalRConnection")] SignalRInvocationContext invocationContext, FunctionContext functionContext)
+        [Function(nameof(OnDisconnected))]
+        public static void OnDisconnected(
+            [SignalRTrigger("chat", "connections", "disconnected", ConnectionStringSetting = "SignalRConnection")]
+                SignalRInvocationContext invocationContext, FunctionContext functionContext)
         {
-            var logger = functionContext.GetLogger("OnDisconnected");
-            logger.LogInformation($"Connection {invocationContext.ConnectionId} is disconnected. Error: {invocationContext.Error}");
+            var logger = functionContext.GetLogger(nameof(OnDisconnected));
+            logger.LogInformation("Connection {connectionId} is disconnected. Error: {error}", invocationContext.ConnectionId, invocationContext.Error);
         }
         // </snippet_on_disconnected>
 
         // <snippet_on_message>
-        [Function("OnClientMessage")]
-        public static void OnClientMessage([SignalRTrigger("Hub", "messages", "sendMessage", "content", ConnectionStringSetting = "SignalRConnection")] SignalRInvocationContext invocationContext, string content, FunctionContext functionContext)
+        [Function(nameof(OnClientMessage))]
+        public static void OnClientMessage(
+            [SignalRTrigger("Hub", "messages", "sendMessage", "content", ConnectionStringSetting = "SignalRConnection")]
+                SignalRInvocationContext invocationContext, string content, FunctionContext functionContext)
         {
-            var logger = functionContext.GetLogger("OnClientMessage");
-            logger.LogInformation($"Connection {invocationContext.ConnectionId} sent a message. Message content: {content}");
+            var logger = functionContext.GetLogger(nameof(OnClientMessage));
+            logger.LogInformation("Connection {connectionId} sent a message. Message content: {content}", invocationContext.ConnectionId, content);
         }
         // </snippet_on_message>
     }