+++ extensions/Worker.Extensions.CosmosDB/release_notes.md
@@ -4,6 +4,6 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB 4.4.1
+### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB <version>
 
-- [Bug] Update CosmosOptions from using IOptionsSnapshot to IOptionsMonitor to avoid DI scoping issues
+- <entry>